HU11 Changes The names of the classes were changed so that their purpose is better understood and the corresponding methods were organized
