--- src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/controller/OrderRestController.java
@@ -1,34 +1,0 @@
-package com.pragma.powerup.usermicroservice.adapters.driving.http.controller;
-
-import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.OrderRequestDto;
-import com.pragma.powerup.usermicroservice.adapters.driving.http.handlers.IOrderHandler;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/order")
-@RequiredArgsConstructor
-public class OrderRestController {
-    private final IOrderHandler orderHandler;
-
-    @Operation(summary = "Add a new Order")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "201", description = "Order created", content = @Content),
-            @ApiResponse(responseCode = "409", description = "Order already exists", content = @Content)
-    })
-    @PostMapping("/")
-    public ResponseEntity<Void> saveOrder(@Valid @RequestBody OrderRequestDto orderRequestDto) {
-        orderHandler.saveOrder(orderRequestDto);
-        return new ResponseEntity<>(HttpStatus.CREATED);
-    }
-}
