--- src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/handlers/IAdminHandler.java
@@ -1,9 +1,0 @@
-package com.pragma.powerup.usermicroservice.adapters.driving.http.handlers;
-
-import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.RestaurantRequestDto;
-import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response.RestaurantResponseDto;
-
-public interface IAdminHandler {
-    void saveRestaurant(RestaurantRequestDto restaurantRequestDto);
-    RestaurantResponseDto getRestaurant(Long id);
-}
