+++ src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/handlers/IDishHandler.java
@@ -4,11 +4,16 @@ package com.pragma.powerup.usermicroservice.adapters.driving.http.handlers;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.DishRequestDto;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.DishUpdateRequest;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.UpdateDishStateRequestDto;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response.CategoryDishesResponseDto;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response.DishResponseDto;
 
-public interface IOwnerHandler {
+import java.util.List;
+
+public interface IDishHandler {
     void saveDish(DishRequestDto dishRequestDto, String idOwner);
     DishResponseDto getDish(Long id);
     void updateDish(DishUpdateRequest dishUpdateRequest,String idOwner);
     void updateState(UpdateDishStateRequestDto updateDishStateRequestDto, String idOwner);
+    List<CategoryDishesResponseDto> getDishesCategorizedByRestaurant(String idRestaurant, int page, int elementsXpage);
+
 }
