+++ src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/handlers/IRestaurantHandler.java
@@ -1,13 +1,14 @@
 package com.pragma.powerup.usermicroservice.adapters.driving.http.handlers;
 
-import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.OrderDishRequestDto;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.RestaurantRequestDto;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response.CategoryDishesResponseDto;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response.ListRestaurantForClientResponseDto;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response.RestaurantResponseDto;
 
 import java.util.List;
 
-public interface IClientHandler {
+public interface IRestaurantHandler {
+    void saveRestaurant(RestaurantRequestDto restaurantRequestDto);
+    RestaurantResponseDto getRestaurant(Long id);
     List<ListRestaurantForClientResponseDto> listRestaurant(int page, int numberOfElements);
-    List<CategoryDishesResponseDto> getDishesCategorizedByRestaurant(String idRestaurant, int page, int elementsXpage);
-    void newOrder(String idRestaurant,String idClient, List<OrderDishRequestDto> ordersDishesModels);
 }
