--- src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/handlers/impl/AdminHandler.java
@@ -1,32 +1,0 @@
-package com.pragma.powerup.usermicroservice.adapters.driving.http.handlers.impl;
-
-import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.RestaurantRequestDto;
-import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response.RestaurantResponseDto;
-import com.pragma.powerup.usermicroservice.adapters.driving.http.handlers.IAdminHandler;
-import com.pragma.powerup.usermicroservice.adapters.driving.http.mapper.IRestaurantRequestMapper;
-import com.pragma.powerup.usermicroservice.adapters.driving.http.mapper.IRestaurantResponseMapper;
-import com.pragma.powerup.usermicroservice.domain.api.IAdminServicePort;
-import com.pragma.powerup.usermicroservice.domain.model.RestaurantModel;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-@Transactional
-public class AdminHandler implements IAdminHandler {
-    private final IAdminServicePort restaurantServicePort;
-    private final IRestaurantRequestMapper restaurantRequestMapper;
-    private final IRestaurantResponseMapper restaurantResponseMapper;
-    @Override
-    public void saveRestaurant(RestaurantRequestDto restaurantRequestDto) {
-        RestaurantModel restaurantModel = restaurantRequestMapper.toRestaurantRequest(restaurantRequestDto);
-        restaurantServicePort.saveRestaurant(restaurantModel);
-    }
-
-    @Override
-    public RestaurantResponseDto getRestaurant(Long id) {
-        RestaurantModel restaurantModel = restaurantServicePort.getRestaurant(id);
-        return restaurantResponseMapper.toResponseRestaurant(restaurantModel);
-    }
-}
