+++ src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/handlers/impl/DishHandler.java
@@ -4,63 +4,81 @@ import com.pragma.powerup.usermicroservice.adapters.driven.client.UserClient;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.DishRequestDto;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.DishUpdateRequest;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.UpdateDishStateRequestDto;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response.CategoryDishesResponseDto;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response.DishResponseDto;
-import com.pragma.powerup.usermicroservice.adapters.driving.http.handlers.IOwnerHandler;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.handlers.IDishHandler;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.mapper.IDishRequestMapper;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.mapper.IDishResponseMapper;
-import com.pragma.powerup.usermicroservice.domain.api.IOwnerServicePort;
-import com.pragma.powerup.usermicroservice.domain.api.IAdminServicePort;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.mapper.IListDishesCategoryByRestaurantMapper;
+import com.pragma.powerup.usermicroservice.domain.api.IDishServicePort;
+import com.pragma.powerup.usermicroservice.domain.api.IRestaurantServicePort;
+import com.pragma.powerup.usermicroservice.domain.exceptions.RestaurantNotExist;
 import com.pragma.powerup.usermicroservice.domain.exceptions.SameStateException;
+import com.pragma.powerup.usermicroservice.domain.model.CategoryWithDishesModel;
 import com.pragma.powerup.usermicroservice.domain.model.DishModel;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.apache.logging.log4j.util.Strings;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
 
 @Service
 @RequiredArgsConstructor
 @Transactional
-public class OwnerHandler implements IOwnerHandler {
-    private final IOwnerServicePort ownerServicePort;
+public class DishHandler implements IDishHandler {
+    private final IDishServicePort dishServicePort;
     private final IDishRequestMapper dishRequestMapper;
     private final IDishResponseMapper dishResponseMapper;
-    private final IAdminServicePort restaurantServicePort;
-    private final UserClient userClient;
+    private final IListDishesCategoryByRestaurantMapper listDishesCategoryByRestaurantMapper;
     @Override
     public void saveDish(DishRequestDto dishRequestDto, String idOwner) {
         DishModel dishModel = dishRequestMapper.toDishRequest(dishRequestDto);
         dishModel.setState(true);
-        ownerServicePort.saveDish(dishModel, idOwner);
+        dishServicePort.saveDish(dishModel, idOwner);
     }
 
     @Override
     public DishResponseDto getDish(Long id) {
-        DishModel dish = ownerServicePort.getDish(id);
+        DishModel dish = dishServicePort.getDish(id);
         return dishResponseMapper.toResponseDish(dish);
     }
 
     @Override
     public void updateDish(DishUpdateRequest dishUpdateRequest,String idOwner) {
-        DishModel dish = ownerServicePort.getDish(dishUpdateRequest.getId());
+        DishModel dish = dishServicePort.getDish(dishUpdateRequest.getId());
 
         if(Strings.isNotBlank(dishUpdateRequest.getDescription()) || Strings.isNotEmpty(dishUpdateRequest.getDescription())) dish.setDescription(dishUpdateRequest.getDescription());
         if(dishUpdateRequest.getPrice() > 0) dish.setPrice(dishUpdateRequest.getPrice());
 
-        ownerServicePort.updateDish(dish, idOwner);
+        dishServicePort.updateDish(dish, idOwner);
     }
 
     @Override
     public void updateState(UpdateDishStateRequestDto dishUpdateStateRequestDto, String idOwner) {
-        DishModel dish = ownerServicePort.getDish(dishUpdateStateRequestDto.getDishId());
+        DishModel dish = dishServicePort.getDish(dishUpdateStateRequestDto.getDishId());
 
         if (dishUpdateStateRequestDto.isState() && dish.getState()) {
             throw new SameStateException();
         }
         dish.setState(dishUpdateStateRequestDto.isState());
 
-        ownerServicePort.updateDishState(dish, idOwner);
+        dishServicePort.updateDishState(dish, idOwner);
     }
 
+    @Override
+    public List<CategoryDishesResponseDto> getDishesCategorizedByRestaurant(String idRestaurant, int page, int elementsXpage) {
+        List<CategoryWithDishesModel> categoryWithDishes;
 
+        try {
+            categoryWithDishes = dishServicePort.getDishesCategorizedByRestaurant(idRestaurant, page, elementsXpage);
+        } catch (RestaurantNotExist e) {
+            throw new RestaurantNotExist();
+        }
+        return categoryWithDishes.stream()
+                .map(listDishesCategoryByRestaurantMapper::toDto)
+                .collect(Collectors.toList());
+    }
 }
