+++ src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/handlers/impl/OrderHandler.java
@@ -1,23 +1,29 @@
 package com.pragma.powerup.usermicroservice.adapters.driving.http.handlers.impl;
 
-import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.OrderRequestDto;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.OrderDishRequestDto;
 import com.pragma.powerup.usermicroservice.adapters.driving.http.handlers.IOrderHandler;
-import com.pragma.powerup.usermicroservice.adapters.driving.http.mapper.IOrderRequestMapper;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.mapper.IOrderDishRequestMapper;
 import com.pragma.powerup.usermicroservice.domain.api.IOrderServicePort;
-import com.pragma.powerup.usermicroservice.domain.model.OrderModel;
+import com.pragma.powerup.usermicroservice.domain.model.OrdersDishesModel;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 @Transactional
 public class OrderHandler implements IOrderHandler {
-    private final IOrderServicePort orderServicePort;
-    private final IOrderRequestMapper orderRequestMapper;
+    private final IOrderServicePort clientServicePort;
+    private final IOrderDishRequestMapper orderDishRequestMapper;
+
     @Override
-    public void saveOrder(OrderRequestDto orderRequestDto) {
-        OrderModel orderModel = orderRequestMapper.toOrderRequest(orderRequestDto);
-        orderServicePort.saveOrder(orderModel);
+    public void newOrder(String idRestaurant, String idClient, List<OrderDishRequestDto> ordersDishesModels) {
+        List<OrdersDishesModel> orderDishes = new ArrayList<>();
+        ordersDishesModels.forEach(dish -> orderDishes.add(orderDishRequestMapper.toModel(dish)));
+
+        clientServicePort.newOrder(idRestaurant, idClient, orderDishes);
     }
 }
