+++ src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/handlers/impl/RestaurantHandler.java
@@ -1,0 +1,50 @@
+package com.pragma.powerup.usermicroservice.adapters.driving.http.handlers.impl;
+
+import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.RestaurantRequestDto;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response.CategoryDishesResponseDto;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response.ListRestaurantForClientResponseDto;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response.RestaurantResponseDto;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.handlers.IRestaurantHandler;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.mapper.IListDishesCategoryByRestaurantMapper;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.mapper.IListRestaurantClientMapper;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.mapper.IRestaurantRequestMapper;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.mapper.IRestaurantResponseMapper;
+import com.pragma.powerup.usermicroservice.domain.api.IOrderServicePort;
+import com.pragma.powerup.usermicroservice.domain.api.IRestaurantServicePort;
+import com.pragma.powerup.usermicroservice.domain.exceptions.RestaurantNotExist;
+import com.pragma.powerup.usermicroservice.domain.model.CategoryWithDishesModel;
+import com.pragma.powerup.usermicroservice.domain.model.RestaurantModel;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+@RequiredArgsConstructor
+@Transactional
+public class RestaurantHandler implements IRestaurantHandler {
+    private final IRestaurantServicePort restaurantServicePort;
+    private final IRestaurantRequestMapper restaurantRequestMapper;
+    private final IRestaurantResponseMapper restaurantResponseMapper;
+    private final IListRestaurantClientMapper listRestaurantClientMapper;
+    @Override
+    public void saveRestaurant(RestaurantRequestDto restaurantRequestDto) {
+        RestaurantModel restaurantModel = restaurantRequestMapper.toRestaurantRequest(restaurantRequestDto);
+        restaurantServicePort.saveRestaurant(restaurantModel);
+    }
+
+    @Override
+    public RestaurantResponseDto getRestaurant(Long id) {
+        RestaurantModel restaurantModel = restaurantServicePort.getRestaurant(id);
+        return restaurantResponseMapper.toResponseRestaurant(restaurantModel);
+    }
+
+    @Override
+    public List<ListRestaurantForClientResponseDto> listRestaurant(int page, int numberOfElements) {
+        return restaurantServicePort.listRestaurant(page, numberOfElements).stream()
+                .map(listRestaurantClientMapper::toDto)
+                .collect(Collectors.toList());
+    }
+}
