+++ src/main/java/com/pragma/powerup/usermicroservice/configuration/BeanConfiguration.java
@@ -12,17 +12,15 @@ import com.pragma.powerup.usermicroservice.adapters.driven.jpa.mysql.repositorie
 import com.pragma.powerup.usermicroservice.adapters.driven.jpa.mysql.repositories.IOrderDishRepository;
 import com.pragma.powerup.usermicroservice.adapters.driven.jpa.mysql.repositories.IOrderRepository;
 import com.pragma.powerup.usermicroservice.adapters.driven.jpa.mysql.repositories.IRestaurantRepository;
-import com.pragma.powerup.usermicroservice.domain.api.IClientServicePort;
-import com.pragma.powerup.usermicroservice.domain.api.IOwnerServicePort;
 import com.pragma.powerup.usermicroservice.domain.api.IOrderServicePort;
-import com.pragma.powerup.usermicroservice.domain.api.IAdminServicePort;
+import com.pragma.powerup.usermicroservice.domain.api.IDishServicePort;
+import com.pragma.powerup.usermicroservice.domain.api.IRestaurantServicePort;
 import com.pragma.powerup.usermicroservice.domain.ports.IDishPersistencePort;
 import com.pragma.powerup.usermicroservice.domain.ports.IOrderPersistencePort;
 import com.pragma.powerup.usermicroservice.domain.ports.IRestaurantPersistencePort;
-import com.pragma.powerup.usermicroservice.domain.usecase.ClientUseCase;
 import com.pragma.powerup.usermicroservice.domain.usecase.OrderUseCase;
-import com.pragma.powerup.usermicroservice.domain.usecase.OwnerUseCase;
-import com.pragma.powerup.usermicroservice.domain.usecase.AdminUseCase;
+import com.pragma.powerup.usermicroservice.domain.usecase.DishUseCase;
+import com.pragma.powerup.usermicroservice.domain.usecase.RestaurantUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -37,8 +35,8 @@ public class BeanConfiguration {
     }
 
     @Bean
-    public IAdminServicePort restaurantServicePort(IRestaurantPersistencePort restaurantPersistencePort, UserClient userClient){
-        return new AdminUseCase(restaurantPersistencePort, userClient);
+    public IRestaurantServicePort restaurantServicePort(IRestaurantPersistencePort restaurantPersistencePort, UserClient userClient){
+        return new RestaurantUseCase(restaurantPersistencePort, userClient);
     }
 
     // Order
@@ -48,8 +46,8 @@ public class BeanConfiguration {
         return new OrderJpaAdapter(orderRepository, orderEntityMapper, orderDishRepository, orderDishEntityMapper);
     }
     @Bean
-    public IOrderServicePort orderServicePort(IOrderPersistencePort orderPersistencePort) {
-        return new OrderUseCase(orderPersistencePort);
+    public IOrderServicePort orderServicePort(IRestaurantPersistencePort restaurantPersistencePort, IDishPersistencePort dishPersistencePort,IOrderPersistencePort orderPersistencePort) {
+        return new OrderUseCase(restaurantPersistencePort, dishPersistencePort, orderPersistencePort);
     }
     // Dish
     @Bean
@@ -57,12 +55,12 @@ public class BeanConfiguration {
         return new DishJpaAdapter(dishRepository, dishEntityMapper);
     }
     @Bean
-    public IOwnerServicePort dishServicePort(IDishPersistencePort dishPersistencePort, IAdminServicePort restaurantServicePort) {
-        return new OwnerUseCase(dishPersistencePort, restaurantServicePort);
+    public IDishServicePort dishServicePort(IDishPersistencePort dishPersistencePort, IRestaurantPersistencePort restaurantPersistencePort, IRestaurantServicePort restaurantServicePort) {
+        return new DishUseCase(dishPersistencePort, restaurantPersistencePort, restaurantServicePort);
     }
 
     @Bean
-    public IClientServicePort clientServicePort(IRestaurantPersistencePort restaurantPersistencePort, IDishPersistencePort dishPersistencePort, IOrderPersistencePort orderPersistencePort) {
-        return new ClientUseCase(restaurantPersistencePort, dishPersistencePort, orderPersistencePort);
+    public IOrderServicePort clientServicePort(IRestaurantPersistencePort restaurantPersistencePort, IDishPersistencePort dishPersistencePort, IOrderPersistencePort orderPersistencePort) {
+        return new OrderUseCase(restaurantPersistencePort, dishPersistencePort, orderPersistencePort);
     }
 }
