+++ src/main/java/com/pragma/powerup/usermicroservice/domain/api/IDishServicePort.java
@@ -1,11 +1,16 @@
 package com.pragma.powerup.usermicroservice.domain.api;
 
 
+import com.pragma.powerup.usermicroservice.domain.model.CategoryWithDishesModel;
 import com.pragma.powerup.usermicroservice.domain.model.DishModel;
 
-public interface IOwnerServicePort {
+import java.util.List;
+
+public interface IDishServicePort {
     void saveDish(DishModel dishModel, String idOwner);
     DishModel getDish(Long id);
     void updateDish(DishModel dishModel,String idOwner);
     void updateDishState(DishModel dishModel, String idOwner);
+    List<CategoryWithDishesModel> getDishesCategorizedByRestaurant(String idRestaurant, int page, int elementsXpage);
+
 }
