+++ src/main/java/com/pragma/powerup/usermicroservice/domain/api/IRestaurantServicePort.java
@@ -1,9 +1,13 @@
 package com.pragma.powerup.usermicroservice.domain.api;
 
+import com.pragma.powerup.usermicroservice.domain.model.CategoryWithDishesModel;
 import com.pragma.powerup.usermicroservice.domain.model.RestaurantModel;
 
-public interface IAdminServicePort {
+import java.util.List;
+
+public interface IRestaurantServicePort {
     void saveRestaurant(RestaurantModel restaurantModel);
     RestaurantModel getRestaurant(Long id);
+    List<RestaurantModel> listRestaurant(int page, int numberOfElements);
 
 }
