--- src/main/java/com/pragma/powerup/usermicroservice/domain/usecase/OwnerUseCase.java
@@ -1,58 +1,0 @@
-package com.pragma.powerup.usermicroservice.domain.usecase;
-
-
-import com.pragma.powerup.usermicroservice.domain.api.IOwnerServicePort;
-import com.pragma.powerup.usermicroservice.domain.api.IAdminServicePort;
-import com.pragma.powerup.usermicroservice.domain.exceptions.UserNotIsOwner;
-import com.pragma.powerup.usermicroservice.domain.model.DishModel;
-import com.pragma.powerup.usermicroservice.domain.model.RestaurantModel;
-import com.pragma.powerup.usermicroservice.domain.ports.IDishPersistencePort;
-
-
-public class OwnerUseCase implements IOwnerServicePort {
-    private final IDishPersistencePort dishPersistencePort;
-    private final IAdminServicePort restaurantServicePort;
-
-    public OwnerUseCase(IDishPersistencePort dishPersistencePort, IAdminServicePort restaurantServicePort) {
-        this.dishPersistencePort = dishPersistencePort;
-        this.restaurantServicePort = restaurantServicePort;
-    }
-
-    @Override
-    public void saveDish(DishModel dishModel, String idOwner) {
-
-        RestaurantModel restaurantModel = restaurantServicePort.getRestaurant(dishModel.getRestaurant().getId());
-
-            if (!restaurantModel.getIdOwner().equals(idOwner)) {
-                throw new UserNotIsOwner();
-            }
-            dishPersistencePort.saveDish(dishModel);
-    }
-
-    @Override
-    public DishModel getDish(Long id) {
-        return dishPersistencePort.getDish(id);
-    }
-
-    @Override
-    public void updateDish(DishModel dishModel,String idOwner) {
-        RestaurantModel restaurantModel = restaurantServicePort.getRestaurant(dishModel.getRestaurant().getId());
-
-        if (!restaurantModel.getIdOwner().equals(idOwner)) {
-            throw new UserNotIsOwner();
-        }
-        dishPersistencePort.updateDish(dishModel);
-    }
-
-    @Override
-    public void updateDishState(DishModel dishModel, String idOwner) {
-        RestaurantModel restaurantModel = restaurantServicePort.getRestaurant(dishModel.getRestaurant().getId());
-
-        if (!restaurantModel.getIdOwner().equals(idOwner)) {
-            throw new UserNotIsOwner();
-        }
-
-        dishPersistencePort.updateSate(dishModel);
-    }
-
-}
