+++ src/test/java/com/pragma/powerup/usermicroservice/domain/usecase/RestaurantUseCaseTest.java
@@ -3,11 +3,16 @@ package com.pragma.powerup.usermicroservice.domain.usecase;
 import com.pragma.powerup.usermicroservice.adapters.driven.client.UserClient;
 import com.pragma.powerup.usermicroservice.adapters.driven.client.feignModels.User;
 import com.pragma.powerup.usermicroservice.domain.model.RestaurantModel;
+import com.pragma.powerup.usermicroservice.domain.ports.IDishPersistencePort;
 import com.pragma.powerup.usermicroservice.domain.ports.IRestaurantPersistencePort;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
+
+import java.util.Collections;
+import java.util.List;
 
 import static org.mockito.Mockito.*;
 
@@ -17,13 +22,15 @@ class RestaurantUseCaseTest {
     private IRestaurantPersistencePort restaurantPersistencePort;
     @Mock
     private UserClient userClient;
+    @Mock
+    IDishPersistencePort dishPersistencePort;
 
-    private AdminUseCase adminUseCase;
+    private RestaurantUseCase restaurantUseCase;
 
     @BeforeEach
     public void setup() {
         MockitoAnnotations.openMocks(this);
-        adminUseCase = new AdminUseCase(restaurantPersistencePort, userClient);
+        restaurantUseCase = new RestaurantUseCase(restaurantPersistencePort, userClient);
     }
 
     @Test
@@ -36,8 +43,18 @@ class RestaurantUseCaseTest {
 
         when(userClient.getUser("1")).thenReturn(user);
 
-        adminUseCase.saveRestaurant(restaurantModel);
+        restaurantUseCase.saveRestaurant(restaurantModel);
 
         verify(restaurantPersistencePort, times(1)).saveRestaurant(restaurantModel);
+    }
+
+    @Test
+    void listRestaurant() {
+        RestaurantModel restaurantModel = DomainData.obtainRestaurant();
+
+        when(restaurantPersistencePort.listByPageAndElements(0, 2))
+                .thenReturn(Collections.singletonList(restaurantModel));
+
+        Assertions.assertInstanceOf(List.class, restaurantUseCase.listRestaurant(0, 2));
     }
 }
